Expose the dvd_stat_t struct.
The commits that brought in the DVDFileStat() function from libdvdread
0.9.7 incorrectly made the stat struct opaque. This can't be done because
the API does not use any allocation or deallocation code. So callers of
DVDFileStat cannot declare stat structs. Since we are attempting to
maintain the API compatibility w/ those releases of libdvdread, the
struct has been brought into the header. Thanks again to Rathann for
bringing this issue to the dvdnav list. And thanks to the original
bug reportera(O. Rolland) to fedora.

git-svn-id: svn://svn.mplayerhq.hu/dvdnav/trunk/libdvdread/src@1188 625a9de0-492d-0410-920c-a7e632af1907